Solo elimine la clase TBModeloTabla porque no era necesaria
